style: bridge into transaction
