+++ common/protocol/filter/tag.go
@@ -3,7 +3,6 @@ package filter
 const (
 	TagTransaction string = "transaction"
 	TagExchange    string = "exchange"
-	TagBridge      string = "bridge"
 	TagCollectible string = "collectible"
 	TagSocial      string = "social"
 	TagDonation    string = "donation"
@@ -13,7 +12,6 @@ const (
 var ValidTag = map[string]string{
 	TagTransaction: TagTransaction,
 	TagExchange:    TagExchange,
-	TagBridge:      TagBridge,
 	TagCollectible: TagCollectible,
 	TagSocial:      TagSocial,
 	TagDonation:    TagDonation,
@@ -23,7 +21,6 @@ var ValidTag = map[string]string{
 var TagPriority = map[string]int{
 	TagTransaction: 1,
 	TagExchange:    2,
-	TagBridge:      2,
 	TagCollectible: 3,
 	TagSocial:      4,
 	TagDonation:    4,
